Fixed route issues
