+++ app/Http/Controllers/ProfileController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -74,6 +75,48 @@ class ProfileController extends Controller
                 'error' => $e->getMessage(),
             ], 500);
         }
+    }
+
+    public function changePassword(Request $request){
+        $user = auth()->user();
+        $fields = Validator::make($request->all(), [
+            'current_password' => 'required|string',
+            'new_password' => 'required|string',
+            'confirm_password' => 'required|string'
+        ]);
+
+        if($fields->fails()){
+            return response()->json([
+                'message' => $fields->messages(),
+                'statusCode' => Response::HTTP_UNPROCESSABLE_ENTITY,
+            ], Response::HTTP_UNPROCESSABLE_ENTITY); // 422
+        }
+
+        if($request->new_password !== $request->confirm_password)
+            return response()->json([
+                'status_code' => Response::HTTP_BAD_REQUEST,
+                'status' => 'error',
+                'message' => 'Passwords do not match',
+            ], Response::HTTP_BAD_REQUEST); // 400
+
+        if (!Hash::check($request->current_password, $user->password_hash)) {
+            return response()->json([
+                'status_code' => Response::HTTP_UNAUTHORIZED,
+                'status' => 'error',
+                'message' => 'Current password is incorrect',
+            ], Response::HTTP_UNAUTHORIZED); // 403
+        }        
+
+        $password= Hash::make($request->new_password);
+        $response = $user->update([
+            'password_hash' => $password
+        ]);
+
+        return response()->json([
+            "message" => "Password changed successfully",
+            "statusCode" => Response::HTTP_OK, // 200
+            "data" => $user
+        ], Response::HTTP_OK);
     }
 
 }
