+++ routes/api.php
@@ -30,7 +30,7 @@ Route::middleware('auth:api')->prefix('v1')->group(function () {
 
     Route::get('/user/search/{nameOrEmail}', [UserController::class, 'search'])->name('search.search');
     Route::get('/user/profile', [ProfileController::class, 'index'])->name('user_profile.index');
-    Route::post('/auth/user/change-password', [LoginController::class, 'changePassword'])->name('user.changePassword');
+    Route::post('/auth/user/change-password', [ProfileController::class, 'changePassword'])->name('user.changePassword');
     Route::patch('/user/bank', [BankDetailController::class, 'addBankDetails'])->name('user.addBankDetails');
     Route::get('/user/all', [UserController::class, 'index'])->name('user.index');
